% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildDT.R
\name{buildDT}
\alias{buildDT}
\title{Build DT::datatable objects from matrix of fit statistics}
\usage{
buildDT(hlme.metrics)
}
\arguments{
\item{hlme.metrics}{A data frame or matrix of model fit metrics for different
number of assumed clusters (G)}
}
\description{
Build DT::datatable objects from matrix of fit statistics
}
