% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{compareClustering}
\alias{compareClustering}
\title{Create co-clustering matrix for FCAL and CRP models}
\usage{
compareClustering(G, models.fcal, models.crp, cutoff = FALSE, threshold = 0.8)
}
\arguments{
\item{G}{Number of clusters assumed by both models}

\item{models.fcal}{List holding all \code{hlme} objects created from FCAL
data}

\item{models.crp}{List holding all \code{hlme} objects created from CRP
data}

\item{cutoff}{Logical, should only subjects assigned posterior cluster
membership probabilities above a given threshold for both models be
included?}

\item{threshold}{Numeric. The posterior probability of cluster membership
cutoff required for a subject to be included in the plot. Only used if
\code{cutoff = TRUE}}
}
\description{
Create co-clustering matrix for FCAL and CRP models
}
