% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_cat}
\alias{describe_cat}
\title{Produce descriptive analysis of categorical variables}
\usage{
describe_cat(vars, data)
}
\arguments{
\item{vars}{Character vector giving the names of categorical variables}

\item{data}{Data frame (or tibble) holding the data of interest}
}
\description{
Produce descriptive analysis of categorical variables
}
