% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{lcmmHeatmap}
\alias{lcmmHeatmap}
\title{Produce a heatmap of LCMM cluster assignments across multiple lcmm::hlme fit objects}
\usage{
lcmmHeatmap(
  models,
  cores = NULL,
  png = TRUE,
  pdf = FALSE,
  file.name = "heatmap",
  return.data = FALSE
)
}
\arguments{
\item{models}{list containing objects generated by
\code{\link[lcmm:hlme]{lcmm::hlme}}.}

\item{cores}{Numeric. Number of CPU cores to use. Defaults to maximum
available.}

\item{png}{Logical. If \code{TRUE}, the heatmap will be saved to file as a
PNG.}

\item{pdf}{Logical. If \code{TRUE}, the heatmap will be saved to file as a
PDF.}

\item{file.name}{Character. File name (without extension) for saved heatmap
files. Defaults to \code{"heatmap"}. Not used if \code{png} and \code{pdf}
are both \code{FALSE}.}

\item{return.data}{Logical. If \code{TRUE}, the averaged co-cluster matrix
will be returned.}
}
\description{
Produce a heatmap of LCMM cluster assignments across multiple lcmm::hlme fit objects
}
