% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetrics.R
\name{makeMetrics}
\alias{makeMetrics}
\title{Produce table of model fit metrics (AIC, BIC, Log-likelihood) for LCMMs}
\usage{
makeMetrics(Gs, models)
}
\arguments{
\item{Gs}{Vector indicating which values of G should metrics be calculated
for.}

\item{models}{List containing LCMM model fits}
}
\description{
Produce table of model fit metrics (AIC, BIC, Log-likelihood) for LCMMs
}
