% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{mlrPlot}
\alias{mlrPlot}
\title{Forest plot of multinomial logistic regression model}
\usage{
mlrPlot(dat, var, class = "class_combined")
}
\arguments{
\item{dat}{Data frame holding covariate and cluster assignment data.}

\item{var}{Character. The name of the covariate(s) of interest. If given as
a character vector then a multivariate model is fitted using the specified
covariates.}

\item{class}{Character. The name of the class assignment variable. Assumed
to be \code{"class_combined"} if not manually specified.}
}
\description{
Forest plot of multinomial logistic regression model
}
