% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankCumulative.R
\name{rankCumulative}
\alias{rankCumulative}
\title{Rank clusters by the area under their mean trajectories}
\usage{
rankCumulative(model, tmax = 7, var.time = "calpro_time")
}
\arguments{
\item{model}{An \code{\link[lcmm]{hlme}} object}

\item{tmax}{Numeric. Maximum observation period. Defaults to 10}

\item{var.time}{Character. Name of variable used for time. Either
"calpro_time" or "crp_time" Also used to determine if the LCMM has been
fitted for FCAL or CRP.}
}
\value{
A data frame with columns for the original cluster number (Original), the new rank (New), and the area calculated via the trapezoidal rule (Area).
}
\description{
Rank clusters by the area under their mean trajectories
}
