% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reLabel.R
\name{reLabel}
\alias{reLabel}
\title{relabel alluvial class labels ensuring consistency.}
\usage{
reLabel(new.G, alluvial.df)
}
\arguments{
\item{new.G}{The number of classes assumed by the LCMM which is now subject
to class relabelling.}

\item{alluvial.df}{Data frame containing class assignments for subjects for
differing G}
}
\description{
Relabel class labels in an alluvial data frame to ensure class
  labels are consistent with the class labels used models assuming fewer
  classes
}
