% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spaghetti.R
\name{spaghettiPlot}
\alias{spaghettiPlot}
\title{Spaghetti plots of each class}
\usage{
spaghettiPlot(
  data,
  models,
  G,
  log = TRUE,
  indi = FALSE,
  multi = TRUE,
  tmax = 10,
  ylim = NULL,
  column = FALSE,
  pprob.cutoff = NA,
  sizes = FALSE,
  mapping = NULL,
  save = FALSE,
  knots = FALSE,
  knot.type = "quantiles",
  n.knots = 3,
  grbf = FALSE,
  l = 1,
  color = NULL,
  clusters = FALSE,
  external = NULL,
  var.time = "calpro_time"
)
}
\arguments{
\item{data}{Data frame used to fit the LCMM(s).}

\item{models}{A list containing \code{\link{hlme}{lcmm}} S3 objects}

\item{G}{Numeric. How many classes does the model assume?}

\item{log}{Logical. Should plots be on log scale? Defaults to TRUE.}

\item{indi}{Logical. Should separate plots for each class be generated?
Defaults to FALSE.}

\item{multi}{Logical. Should sub-plots for each class be plotted alongside
each other? Defaults to TRUE.}

\item{tmax}{Numeric. Maximum observation period. Defaults to 10}

\item{ylim}{Character. What should the y-axis of the plots be scaled to fit?
Either NULL (default), "data", "pred", or "conf". The default uses
pre-determined values for FCAL and CRP}

\item{column}{Logical. Should all sub-plots be in a single column? Defaults
to FALSE (two columns)}

\item{pprob.cutoff}{Posterior probability cut-off for subjects to be included
as trajectories}

\item{sizes}{Logical. Should class sizes be given as plot titles?}

\item{mapping}{Numeric vector Giving the new ordering of sub-plots. One must
take into account plots are generated by column - not row. If not specified
then the default class numbers will be used.}

\item{save}{Logical. Should sub-figure labels be generated?}

\item{knots}{Logical. If TRUE, knot locations will be plotted as vertical
lines. Defaults to FALSE}

\item{knot.type}{Character. How are knots placed? Either "quantiles"
(default) or "equal" (equally spaced). Not used if \code{knots = FALSE}.}

\item{n.knots}{Numeric integer giving the number of knots used. Defaults to
3. Not used if \code{knots = FALSE}.}

\item{grbf}{Logical. If \code{TRUE}, then the model is assumed to have been
fitted using two Gaussian Radial basis functions. Defaults to FALSE (using
natural cubic splines instead).}

\item{l}{Numeric. Length scale constant for the Gaussian radial basis
functions (not used if \code{GRBF = FALSE})}

\item{color}{Character. Name of variable in \code{data} to use to colour
lines and points corresponding to subjects. If NULL then lines and points
will be grey.}

\item{clusters}{Logical. Are clusters given via a "class" column in
\code{data}? Otherwise clusters are found from the \code{hlme} object in
models[[G]]. Defaults to FALSE.}

\item{external}{Optional data frame providing mean cluster
estimates from \code{\link{predictY}{lcmm}}. Intended to be used with
outputs from models applied to an external cohort. Columns should be
"time", "Ypred", "CI1", and "CI2".}

\item{var.time}{Character. Name of variable used for time. Either
"calpro_time" or "crp_time" Also used to determine if the LCMM has been
fitted for FCAL or CRP.}
}
\description{
Spaghetti plots of each class
}
