% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{mlrPlot}
\alias{mlrPlot}
\title{Forest plot of multinomial logistic regression model}
\usage{
mlrPlot(
  dat,
  var,
  class = "class_combined",
  prob = "probmax",
  minprob = 0.5,
  extern = NULL
)
}
\arguments{
\item{dat}{Data frame holding covariate and cluster assignment data.}

\item{var}{Character. The name of the covariate(s) of interest. If given as
a character vector then a multivariate model is fitted using the specified
covariates.}

\item{class}{Character. The name of the class assignment variable. Assumed
to be \code{"class_combined"} if not manually specified.}

\item{prob}{Character. The name of the variable which gives posterior
probabilities for cluster membership. Assumed to be \code{"probmax"} if not
manually specified.}

\item{minprob}{Numeric. The minimum posterior probability for cluster
membership required for a subject to be included in the analysis. Assumed
to be 0.5 if not manually specified}

\item{extern}{Optional data frame containing multinomial logistic
regression results obtained from an external dataset. If not given then it
is assumed that no external data is required.}
}
\description{
Forest plot of multinomial logistic regression model
}
