% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associations.R
\name{plotCat}
\alias{plotCat}
\title{Percentage bar plots for categorical data}
\usage{
plotCat(dat, var)
}
\arguments{
\item{dat}{data frame holding categorical data and cluster assignment data, the latter denoted via a "class_combined" column.}

\item{var}{Character. The name of the categorical variable of interest}
}
\value{
A \code{\link[patchwork]{patchwork}} object
}
\description{
Percentage bar plots for categorical data
}
